Add insert_macro class to Scope
